requirements.txt
numpy
opencv-python
mediapipe
PyOpenGL
glfw
